refactor(browser): split Playwright interactions
